copying the main setup.
