+++ src/MGmolInterface.h
@@ -23,6 +23,7 @@ public:
     virtual int setupLRs(const std::string input_file)                  = 0;
     virtual int setupConstraintsFromInput(const std::string input_file) = 0;
     virtual void run()                                                  = 0;
+    virtual void setup()                                                = 0;
 };
 
 #endif
